test(analysis): consolidate diff table regression coverage (#1138)
